Fix flaws in detecting color support (#13)
* Fix flaws in detecting color support

Consistently using the correct new function name would have been a good
start, but in addition the logic was flawed for Windows.

Cygwin and Msys (and therefore Git Bash), emulate pseudo terminals by
using named pipes, so `sapi_windows_vt100_support` and any test for a
tty will fail. However, this is not the case when Git Bash is given the
name of a program to run that requires a PATH lookup, rather than being
given the full filename.

As such, we should only rely on the specific environment variables.