Fixed failures when PHP code is supplied on stdin
When PHP is interpreting the code from STDIN (as opposed to executing a
script file) it sets $argv[0] to a hardcoded string (specific string
depends on PHP version). XDebugHandler was supplying this string to the
restarted subprocess, which subsequently failed with 'Could not open
input file: ...'.

Now these strings are replaced with '--' and subprocess does not fail.