Now allow a restart with directly executed code
If XdebugHandler is auto-prepended, stdin can be read by the restarted
process. If this is not the case, then the restarted process will have
nothing to execute and will exit immediately if it is not interactive.

Status messages have been added to show if directly executed code is
being used and to report when the restarted process exits. Additionally
the constant STDOUT is checked before use, as it is not defined when
code is provided via stdin.

Thanks again to Bruce Weirdan https://github.com/weirdan