Define API and allow XdebugHandler to be extended
As developers try to incorporate extra functionality it makes sense to
define the API and provide extendable hooks into the core.

These have been restricted to a new requiresRestart method, for extra
control, and by modifying the restart method to allow access to the
temporary ini file using the tmpIni property, which is now protected.