Simplify ini value escaping and avoid using ctype
Using ctype_alnum to determine if an ini value needed to be escaped was
chosen because it replicated -d command-line parsing in PHP source.

However, the internal scanner returns everything as a string, having
either translated the special constants (On, Off, True, False, Yes, No
and None) to a boolean, or removed any outer quotes.

Since we are only dealing with parsed ini values, it is safe (and
faster) to simply double-quote each value.