Make ini value escaping more robust
There is a PHP ini-parsing bug (https://bugs.php.net/bug.php?id=55091)
that chokes on a double-quote newline sequence and could result in the
value being split. This is highly unlikely to be a problem. Furthermore,
it needs to be a valid PHP ini setting that has been `ini_set`. As such
it will be reset again in the restarted process.