Use command-line option to suppress ini scanning
When specifying an ini file on the command-line, there are two ways to
stop PHP scanning for additional ini files - either by setting
`PHP_INI_SCAN_DIR` to an empty string or using the `-n` option. Neither
are documented as doing this in the PHP manual.

xdebug-handler always used the first method (and restored the
environment in the restart). Later, after debugging through the PHP
source code, it became apparent that another solution was to use `-n`,
but there seemed little point in changing something that was already
working at that time.

Note that `main/php_ini.c` is quite explicit about this usage, with the
following code comment, from August 2008:

    If SAPI does not want to ignore all ini files OR an overriding
    file/path is given. This allows disabling scanning for ini files in
    the PHP_CONFIG_FILE_SCAN_DIR but still load an optional ini file.

Switching to this method makes it easier for developers to use the
restart settings to invoke a PHP sub-process, because the environment is
'clean' in that sub-process. This change is internal and has no affect
on current usage.