Check that php_ini_scanned_files() is functional
If the PHP build does not set a --with-config-file-scan-dir value (on
Windows mainly), `php_ini_scanned_files` always fails. This is now
fixed in 7.1.13 and 7.2.1. See https://bugs.php.net/73124

When xdebug-handler used the environment to suppress ini-scanning, this
did not matter - although xdebug may not have been disabled. But now
that command-line options are used, configuration from the scanned inis
will not be used in the restart.

This fix ensures that the restart fails in such a situation, with an
appropriate status message.