Add Windows CTRL event handling on PHP 7.4+
Windows console event handling can be problematic because CTRL+C signals
may be disabled and this is inherited by child processes. Also there is
no `pcntl_signal_get_handler` equivalent in Windows PHP, so it is
impossible to know if signals are enabled or if a handler has been
registered.

When creating a handler, `sapi_windows_set_ctrl_handler` automatically
enables CTRL+C signals in the current process (and by default any child
process). In order to enable and transfer CTRL handling to the restarted
process, we register a handler in the parent that doesn't do anything
resulting in any existing handler being overwritten and the suppression
of CTRL events.