Bypass shell when restarting in PHP 7.4+
In PHP-7.4 `proc_open` accepts an array of command-line options which
are automatically quoted for the OS. This enables the shell to be
bypassed and the php process to be called directly for the restart.

This feature is used if available, otherwise the command-line is quoted
as before. Support for `passthru` has been dropped to make life simpler.